Move modern_forms base entity to separate module
