+++ homeassistant/components/modern_forms/light.py
@@ -17,7 +17,7 @@ from homeassistant.util.percentage import (
     ranged_value_to_percentage,
 )
 
-from . import ModernFormsDeviceEntity, modernforms_exception_handler
+from . import modernforms_exception_handler
 from .const import (
     ATTR_SLEEP_TIME,
     CLEAR_TIMER,
@@ -28,6 +28,7 @@ from .const import (
     SERVICE_SET_LIGHT_SLEEP_TIMER,
 )
 from .coordinator import ModernFormsDataUpdateCoordinator
+from .entity import ModernFormsDeviceEntity
 
 BRIGHTNESS_RANGE = (1, 255)
